search/requirements.txt
fastapi
uvicorn
openai
qdrant-client
python-dotenv
PyYAML
pydantic
mangum